Cortex v0.0.1 Cortex.Thing

Functions

build_module(name)

changeset(model, params \\ :empty)

Creates a changeset based on the model and params.

If no params are provided, an invalid changeset is returned with no validation performed.

get_code_for(name)

module_name(name)