Cortex v0.0.1 Cortex.Thing.Manager

A loop identifies each unknown and unconnected operating system tty that looks like a serial port by probing it (connecting to it) in order to launch stored Firmata code under a supervisor.

If no probes are active, waits 5 seconds before looping again

If probes are active, waits 10 seconds, then closes any open probe

On receive firmware_name, closes the probe and saves the name

For each tty identification, lookup matching firmware_name in database

If code is found in the database for this firmware_name, start it supervised

Detects known devices which have been lost/unplugged and stops them via supervisor

Functions

disconnect(ident_pids, tty_path, found_callback)

takes a list of pid tuples, finds the one with the tty name, removes it from the list, calls the callback with it, and then returns the list

handle_call(msg, from, state)

Causes identification on each unknown and unconnected tty

handle_info(arg1, state)

Received when we have lost a known Firmata device Stopping any process you started when it was identified

identified(ttys, tty_path, firmware_name)

Takes a list of tty tuples and updates the one with the given path with the given name and status 2 triggering a disconnect event for it and returning the list

identify(ttys)

Takes a list of tuples and examines the status of each one. If 0 we trigger a scan and set the status to 1 so we dont trigger again returns the list

loop(manager)

lost_knowns(prev_ttys, new_ttys)

Reveals which ttys that were otherwise known, are now lost

remove_connected(ttys)

Takes list of tty tuples and returns a new list with any that were in the connected state having been removed

start_link()

update_tty_list(sys_ttys, seen_ttys)

Takes the system tty and current list of seen ttys and produces new list of seen ttys by removing or adding while converting to desired tuple structure